fix(project): Extract to components
