+++ src/App.jsx
@@ -1,47 +1,9 @@
 import {useState} from "react";
 import confetti from "canvas-confetti"
+import {TURNS} from "./const/constants.js"
+import {checkWinnerFrom, checkEndGame} from "./logic/board.js"
+import Board from "./components/Board.jsx";
 
-const TURNS = {
-    X: 'x',
-    O: 'o'
-}
-const Square = ({children, isSelected, updateBoard, index}) => {
-    const className = `square ${isSelected ? 'is-selected' : ''}`
-    const handleClick = () => {
-        updateBoard(index)
-    }
-    return (
-        <div onClick={handleClick} className={className}>
-            {children}
-        </div>
-    )
-}
-
-const WINNER_COMBOS = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6]
-]
-
-const checkWinner = (boardToCheck) => {
-    // Aqui revisamos todas las combinaciones ganadoras
-    for (const combo of WINNER_COMBOS) {
-        const [a, b, c] = combo
-        if (
-            boardToCheck[a] &&
-            boardToCheck[a] === boardToCheck[b] &&
-            boardToCheck[a] === boardToCheck[c]
-        ) {
-            return boardToCheck[a]
-        }
-    }
-    return null
-}
 const App = () => {
     const [board, setBoard] = useState(
         Array(9).fill(null)
@@ -58,7 +20,7 @@ const App = () => {
         setBoard(newBoard)
         const newTurn = turn === TURNS.X ? TURNS.O : TURNS.X
         setTurn(newTurn)
-        const newWinner = checkWinner(newBoard)
+        const newWinner = checkWinnerFrom(newBoard)
         if (newWinner) {
             setWinner(newWinner)
             confetti()
@@ -72,53 +34,9 @@ const App = () => {
         setTurn(TURNS.X)
         setWinner(null)
     }
-
-    const checkEndGame = (newBoard) => {
-        return newBoard.every((square) => square !== null)
-    }
-
     return (
-        <main className="board">
-            <h1>Tic Tac Toe</h1>
-            <button onClick={resetGame}>Reset del Juego</button>
-            <section className="game">
-                {
-                    board.map((_, index) => {
-                        return (
-                            <Square key={index} index={index} updateBoard={updateBoard}>
-                                {board[index]}
-                            </Square>
-                        )
-                    })
-                }
-            </section>
-            <section className="turn">
-                <Square isSelected={turn === TURNS.X}>{TURNS.X}</Square>
-                <Square isSelected={turn === TURNS.O}>{TURNS.O}</Square>
-            </section>
-            {
-                winner !== null && (
-                    <section className="winner">
-                        <div className="text">
-                            <h2>
-                                {
-                                    winner === false
-                                        ? 'Empate'
-                                        : 'Ganó'
-                                }
-                            </h2>
-                            <header className="win">
-                                {winner && <Square>{winner}</Square>}
-                            </header>
-                            <footer>
-                                <button onClick={resetGame}>Empezar de nuevo</button>
-                            </footer>
-                        </div>
-                    </section>
-                )
-            }
-        </main>
-    );
+        <Board winner={winner} updateBoard={updateBoard} turn={turn} board={board} resetGame={resetGame} />
+    )
 }
 
 export default App
