+++ src/components/Square.jsx
@@ -1,0 +1,11 @@
+export const Square = ({children, isSelected, updateBoard, index}) => {
+    const className = `square ${isSelected ? 'is-selected' : ''}`
+    const handleClick = () => {
+        updateBoard(index)
+    }
+    return (
+        <div onClick={handleClick} className={className}>
+            {children}
+        </div>
+    )
+}
