+++ src/logic/board.js
@@ -1,0 +1,20 @@
+import {WINNER_COMBOS} from "../const/constants.js";
+
+export const checkWinnerFrom = (boardToCheck) => {
+    // Aqui revisamos todas las combinaciones ganadoras
+    for (const combo of WINNER_COMBOS) {
+        const [a, b, c] = combo
+        if (
+            boardToCheck[a] &&
+            boardToCheck[a] === boardToCheck[b] &&
+            boardToCheck[a] === boardToCheck[c]
+        ) {
+            return boardToCheck[a]
+        }
+    }
+    return null
+}
+
+export const checkEndGame = (newBoard) => {
+    return newBoard.every((square) => square !== null)
+}
